refactor to extract window framer state
